Support build configs

+++ mandelbulber2/tools/bench/bench.bat
@@ -20,12 +20,25 @@ set _SCRIPT_DRIVE=%~d0
 set _SCRIPT_FOLDER=%~dp0
 set timestamp=%yy%%mm%%dd%%hh%%min%%ss%_%dow%
 set SRC=%_SCRIPT_FOLDER%\..\..\..\
-set OutDir=%SRC%\build-win\Release\x64
 set IntDir=%SRC%\build-win\Bench\%timestamp%
+
+REM # Support build configs #
+set ChkDir=%SRC%\build-win\AnalysisRelease\Win32
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
+set ChkDir=%SRC%\build-win\Debug\Win32
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
+set ChkDir=%SRC%\build-win\Release\Win32
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
+set ChkDir=%SRC%\build-win\AnalysisRelease\x64
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
+set ChkDir=%SRC%\build-win\Debug\x64
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
+set ChkDir=%SRC%\build-win\Release\x64
+if exist %ChkDir%\mandelbulber2.exe set OutDir=%ChkDir%
 set bulb=%OutDir%\mandelbulber2.exe
 
 REM # Clean IntDir #
-rd /s /q %IntDir%
+if exist %IntDir%\ rd /s /q %IntDir%
 mkdir %IntDir%
 mkdir %IntDir%\ocl
